# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReineiraOS Modules API",
    "version": "0.1.0",
    "description": "Backend API for ReineiraOS venture applications"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/escrows": {
      "post": {
        "operationId": "createEscrow",
        "tags": [
          "Escrow"
        ],
        "summary": "Create a new escrow",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "minLength": 1
                  },
                  "counterparty": {
                    "type": "string",
                    "minLength": 1
                  },
                  "deadline": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "external_reference": {
                    "type": "string",
                    "minLength": 1
                  },
                  "amount": {
                    "type": "number",
                    "exclusiveMinimum": 0
                  },
                  "currency": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "fiat",
                          "crypto"
                        ]
                      },
                      "code": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "type",
                      "code"
                    ]
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "resolver_address": {
                    "type": "string",
                    "minLength": 1
                  },
                  "insurance": {
                    "type": "object",
                    "properties": {
                      "pool_address": {
                        "type": "string",
                        "minLength": 1
                      },
                      "policy_address": {
                        "type": "string",
                        "minLength": 1
                      },
                      "coverage_amount": {
                        "type": "number",
                        "exclusiveMinimum": 0
                      },
                      "expiry": {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      }
                    }
                  }
                },
                "required": [
                  "type",
                  "amount",
                  "currency"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Escrow created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "public_id": {
                      "type": "string"
                    },
                    "contract_address": {
                      "type": "string"
                    },
                    "abi_function_signature": {
                      "type": "string"
                    },
                    "abi_parameters": {
                      "type": "object",
                      "properties": {
                        "encrypted_owner": {
                          "type": "array",
                          "maxItems": 4,
                          "minItems": 4,
                          "prefixItems": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "string"
                            }
                          ]
                        },
                        "encrypted_amount": {
                          "type": "array",
                          "maxItems": 4,
                          "minItems": 4,
                          "prefixItems": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "number"
                            },
                            {
                              "type": "string"
                            }
                          ]
                        },
                        "resolver": {
                          "type": "string"
                        },
                        "resolver_data": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "encrypted_owner",
                        "encrypted_amount",
                        "resolver",
                        "resolver_data"
                      ]
                    }
                  },
                  "required": [
                    "public_id",
                    "contract_address",
                    "abi_function_signature",
                    "abi_parameters"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "statusCode"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "statusCode"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listEscrows",
        "tags": [
          "Escrow"
        ],
        "summary": "List escrows for current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated escrow list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "public_id": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          },
                          "counterparty": {
                            "type": "string"
                          },
                          "deadline": {
                            "type": "string"
                          },
                          "external_reference": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": [
                                  "fiat",
                                  "crypto"
                                ]
                              },
                              "code": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "type",
                              "code"
                            ]
                          },
                          "status": {
                            "type": "string"
                          },
                          "on_chain_id": {
                            "type": "string"
                          },
                          "tx_hash": {
                            "type": "string"
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "resolver_address": {
                            "type": "string"
                          },
                          "pool_address": {
                            "type": "string"
                          },
                          "policy_address": {
                            "type": "string"
                          },
                          "coverage_id": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "public_id",
                          "type",
                          "amount",
                          "currency",
                          "status",
                          "created_at"
                        ]
                      }
                    },
                    "continuation_token": {
                      "type": "string"
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "limit": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "items",
                    "has_more",
                    "limit"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "statusCode"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}